20191114 modify warnings

+++ Everyday/0tobiipro/Tobii4C/GUI/Demo/main.cpp
@@ -34,10 +34,10 @@ void MyEmit::send_emit()
 void gaze_point_callback(tobii_gaze_point_t const *gaze_point, void *user_data) {
     if (gaze_point->validity == TOBII_VALIDITY_VALID){
         printf("Gaze point: %f, %f\n",
-               gaze_point->position_xy[0],
-               gaze_point->position_xy[1]);
-        X=gaze_point->position_xy[0];
-        Y=gaze_point->position_xy[1];
+               static_cast<double>(gaze_point->position_xy[0]),
+               static_cast<double>(gaze_point->position_xy[1]));
+        X=static_cast<double>(gaze_point->position_xy[0]);
+        Y=static_cast<double>(gaze_point->position_xy[1]);
     }
 }
